python/pwntools/pyproject.toml
[project]
name = "pwntools_test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pwntools>=4.14.1",
]
